pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-synth"
version = "0.1.0"
description = "Automated pipeline for synthesizing LLM training data from skill/prompt collections"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "openai>=1.30",
    "anthropic>=0.30",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tiktoken>=0.7",
    "rich>=13.7",
    "xxhash>=3.4",
]

[project.scripts]
skill-synth = "skill_synth.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_synth"]
